// packages/test/generatedCode/models/plannerCategoryDescriptions.ts
import {AdditionalDataHolder, Parsable} from '@microsoft/kiota-abstractions';

export interface PlannerCategoryDescriptions extends Partial<AdditionalDataHolder>, Partial<Parsable> {
    /** Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well. */
    additionalData?: Record<string, unknown>;
    /** The label associated with Category 1 */
    category1?: string;
    /** The label associated with Category 10 */
    category10?: string;
    /** The label associated with Category 11 */
    category11?: string;
    /** The label associated with Category 12 */
    category12?: string;
    /** The label associated with Category 13 */
    category13?: string;
    /** The label associated with Category 14 */
    category14?: string;
    /** The label associated with Category 15 */
    category15?: string;
    /** The label associated with Category 16 */
    category16?: string;
    /** The label associated with Category 17 */
    category17?: string;
    /** The label associated with Category 18 */
    category18?: string;
    /** The label associated with Category 19 */
    category19?: string;
    /** The label associated with Category 2 */
    category2?: string;
    /** The label associated with Category 20 */
    category20?: string;
    /** The label associated with Category 21 */
    category21?: string;
    /** The label associated with Category 22 */
    category22?: string;
    /** The label associated with Category 23 */
    category23?: string;
    /** The label associated with Category 24 */
    category24?: string;
    /** The label associated with Category 25 */
    category25?: string;
    /** The label associated with Category 3 */
    category3?: string;
    /** The label associated with Category 4 */
    category4?: string;
    /** The label associated with Category 5 */
    category5?: string;
    /** The label associated with Category 6 */
    category6?: string;
    /** The label associated with Category 7 */
    category7?: string;
    /** The label associated with Category 8 */
    category8?: string;
    /** The label associated with Category 9 */
    category9?: string;
}